Task 1 Question 1
The program would detect if the start range and end range are positive
integers. if not, the program would display warning messages.

If both ranges are positive integers, then the core of the program will
start. A for loop is used to go through the range of the numbers and
PHP modulus is used to check if the number is either a multiple of 3
which the program would print "Fizz" or if its a multiple of 5, it
would print "Buzz" and if it is not either a multiple of 3 and not
multiple of 5 then it would print the integer itself.

Assumption, if the number is a multiple of 3 and 5 (eg. 15), the script
would display both the Fizz and Buzz together making FizzBuzz