Task 1 Question 2: adding the bazz
To capture when the Fizz and Buzz is being called consecutively,
counter called consecutive is made to count when either the Buzz or the
Fizz has been called consecutively. If the conter reaches 2 means that
the Fizz or Buzz have been called 2 times consecutively, so the next
item would be changed to Bazz. If the Bazz or a normal integer is
printed, then the consecutive counter is reset to 0 to capture the next
consecutive call